+++ order-service/src/main/java/n11bootcamp_project_backend/order_service/dto/event/OrderEventMessage.java
@@ -1,0 +1,8 @@
+package n11bootcamp_project_backend.order_service.dto.event;
+
+import java.util.List;
+
+public record OrderEventMessage(
+        String orderId,
+        List<OrderItemMessage> items
+) {}