+++ order-service/src/main/java/n11bootcamp_project_backend/order_service/dto/event/OrderItemMessage.java
@@ -1,0 +1,6 @@
+package n11bootcamp_project_backend.order_service.dto.event;
+
+public record OrderItemMessage(
+        String productId,
+        Integer quantity
+) {}