+++ stock-service/src/main/java/n11bootcamp_project_backend/stock_service/config/RabbitMQConfig.java
@@ -1,0 +1,59 @@
+package n11bootcamp_project_backend.stock_service.config;
+
+import org.springframework.amqp.core.Binding;
+import org.springframework.amqp.core.BindingBuilder;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.core.TopicExchange;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
+import org.springframework.amqp.support.converter.MessageConverter;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
+
+
+@Configuration
+public class RabbitMQConfig {
+
+    @Bean
+    public TopicExchange sagaExchange() {
+        return new TopicExchange("saga.exchange");
+    }
+
+    @Bean
+    public TopicExchange loggingExchange() {
+        return new TopicExchange("logging.exchange");
+    }
+
+    // order.created eventini dinleyecek queue
+    @Bean
+    public Queue stockOrderQueue() {
+        return new Queue("stock.order.queue", true);
+    }
+
+    @Bean
+    public Binding stockOrderBinding(Queue stockOrderQueue,
+                                     TopicExchange sagaExchange) {
+        return BindingBuilder
+                .bind(stockOrderQueue)
+                .to(sagaExchange)
+                .with("order.created");
+    }
+
+    @Bean
+    public MessageConverter jsonMessageConverter() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        return new Jackson2JsonMessageConverter(objectMapper);
+    }
+
+    @Bean
+    public SimpleRabbitListenerContainerFactory rabbitListenerContainerFactory(
+            ConnectionFactory connectionFactory) {
+        SimpleRabbitListenerContainerFactory factory =
+                new SimpleRabbitListenerContainerFactory();
+        factory.setConnectionFactory(connectionFactory);
+        factory.setMessageConverter(jsonMessageConverter());
+        return factory;
+    }
+}